@@ -35,7 +35,7 @@
     "babel-eslint": "^10.0.1",
     "babel-jest": "^23.4.2",
     "del-cli": "^1.1.0",
-    "eslint": "^5.9.0",
+    "eslint": "^5.11.1",
     "eslint-config-mainframe": "^2.3.0",
     "flow-bin": "^0.89.0",
     "flow-copy-source": "^2.0.2",
@@ -44,7 +44,7 @@
     "jest-fetch-mock": "^2.1.0",
     "jest-junit": "^5.1.0",
     "jest-puppeteer": "^3.5.1",
-    "lerna": "^3.5.1",
+    "lerna": "^3.8.1",
     "prettier": "^1.15.3",
     "puppeteer": "^1.11.0"
   }
